+++ Hera.Data/Infrastructure/EntityBase.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 namespace Hera.Data.Infrastructure
 {
-    public abstract class EntityBase: EntityBaseTypeId<int>
+    public abstract class EntityBase: EntityBaseTypeId<long>
     {
     }
 }
