+++ Hera.Services/ServiceBase.cs
@@ -2,9 +2,9 @@
 
 namespace Hera.Services
 {
-    public class ServiceBase<T> : ServiceBaseTypeId<T, int>, IServiceBase<T> where T : class, IEntityTypeId<int>
+    public class ServiceBase<T> : ServiceBaseTypeId<T, long>, IServiceBase<T> where T : class, IEntityTypeId<long>
     {
-        public ServiceBase(IRepositoryBaseTypeId<T, int> repository) : base(repository)
+        public ServiceBase(IRepositoryBaseTypeId<T, long> repository) : base(repository)
         {
         }
     }
